% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.aipw_curve_one_direction}
\alias{.aipw_curve_one_direction}
\title{AIPW curve slopes for one fold direction using aipw_curve_cpp}
\usage{
.aipw_curve_one_direction(rf, X, Y, honest_idx, var, grid, ghat, subset = NULL)
}
\description{
AIPW curve slopes for one fold direction using aipw_curve_cpp
}
\keyword{internal}
