% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.aipw_effect_binary}
\alias{.aipw_effect_binary}
\title{Full AIPW binary effect with cross-fitting and repeated honest splits}
\usage{
.aipw_effect_binary(object, var, subset = NULL, propensity_trees = 2000L)
}
\description{
Full AIPW binary effect with cross-fitting and repeated honest splits
}
\keyword{internal}
