% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.aipw_one_direction}
\alias{.aipw_one_direction}
\title{AIPW scores for one fold direction using aipw_scores_cpp}
\usage{
.aipw_one_direction(
  rf,
  X,
  Y,
  honest_idx,
  var,
  a,
  b,
  is_binary,
  ghat,
  subset = NULL
)
}
\description{
AIPW scores for one fold direction using aipw_scores_cpp
}
\keyword{internal}
