% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_pred_vector}
\alias{.get_pred_vector}
\title{Extract predictions as numeric vector (handles both regression and probability forests)}
\usage{
.get_pred_vector(rf, newdata)
}
\description{
Extract predictions as numeric vector (handles both regression and probability forests)
}
\keyword{internal}
