% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.get_X_ord}
\alias{.get_X_ord}
\title{Get ranger-ordered X matrix, using cache on infForest object if available}
\usage{
.get_X_ord(object, rf)
}
\description{
Get ranger-ordered X matrix, using cache on infForest object if available
}
\keyword{internal}
