% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.residualize_Y}
\alias{.residualize_Y}
\title{Residualize Y by fitting a leave-variable-out forest on the build fold
Returns a Y vector where honest observations have h_hat subtracted}
\usage{
.residualize_Y(X, Y, build_idx, honest_idx, var)
}
\description{
Residualize Y by fitting a leave-variable-out forest on the build fold
Returns a Y vector where honest observations have h_hat subtracted
}
\keyword{internal}
