% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infForest.R
\name{print.infForest}
\alias{print.infForest}
\title{Print method for infForest objects}
\usage{
\method{print}{infForest}(x, ...)
}
\arguments{
\item{x}{An \code{infForest} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for infForest objects
}
