% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_diagnose.R
\name{print.infForest_ct}
\alias{print.infForest_ct}
\title{Print method for infForest_ct objects}
\usage{
\method{print}{infForest_ct}(x, ...)
}
\arguments{
\item{x}{An \code{infForest_ct} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for infForest_ct objects
}
