% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_curve.R
\name{print.infForest_curve}
\alias{print.infForest_curve}
\title{Print method for infForest_curve objects}
\usage{
\method{print}{infForest_curve}(x, ...)
}
\arguments{
\item{x}{An \code{infForest_curve} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for infForest_curve objects
}
