% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{print.infForest_effect}
\alias{print.infForest_effect}
\title{Print method for infForest_effect objects}
\usage{
\method{print}{infForest_effect}(x, ...)
}
\arguments{
\item{x}{An \code{infForest_effect} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for infForest_effect objects
}
