% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{print.infForest_interaction}
\alias{print.infForest_interaction}
\title{Print method for infForest_interaction objects}
\usage{
\method{print}{infForest_interaction}(x, ...)
}
\arguments{
\item{x}{An \code{infForest_interaction} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for infForest_interaction objects
}
