% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.infForest_summary}
\alias{print.infForest_summary}
\title{Print method for infForest_summary objects}
\usage{
\method{print}{infForest_summary}(x, ...)
}
\arguments{
\item{x}{An \code{infForest_summary} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for infForest_summary objects
}
