% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.aipw_build_curve}
\alias{.aipw_build_curve}
\title{Build AIPW effect curve for continuous predictors (no augmentation)}
\usage{
.aipw_build_curve(
  object,
  var,
  grid_lo,
  grid_hi,
  n_honest,
  bw,
  subset = NULL,
  propensity_trees = 2000L
)
}
\description{
Build AIPW effect curve for continuous predictors (no augmentation)
}
\keyword{internal}
