% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.aipw_one_direction}
\alias{.aipw_one_direction}
\title{Hybrid AIPW scores: within-leaf working model + propensity correction}
\usage{
.aipw_one_direction(
  rf,
  X,
  Y,
  honest_idx,
  var,
  a,
  b,
  is_binary,
  ghat,
  subset = NULL
)
}
\description{
Hybrid AIPW scores: within-leaf working model + propensity correction
}
\keyword{internal}
