% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{.fit_propensity}
\alias{.fit_propensity}
\title{Fit propensity model: predict X_j from X_{-j}}
\usage{
.fit_propensity(X, var, is_binary, n_trees = 2000L)
}
\description{
Fit propensity model: predict X_j from X_{-j}
}
\keyword{internal}
