% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.residualize_FWL}
\alias{.residualize_FWL}
\title{Full nonparametric FWL with cross-fitted nuisance estimation
K-fold cross-fitting ensures each obs's h_hat/g_hat didn't see that obs's Y.}
\usage{
.residualize_FWL(X, Y, build_idx, honest_idx, var, K = 5L)
}
\description{
Full nonparametric FWL with cross-fitted nuisance estimation
K-fold cross-fitting ensures each obs's h_hat/g_hat didn't see that obs's Y.
}
\keyword{internal}
