% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_curve.R
\name{plot.infForest_curve}
\alias{plot.infForest_curve}
\title{Plot method for infForest_curve objects}
\usage{
\method{plot}{infForest_curve}(x, ...)
}
\arguments{
\item{x}{An \code{infForest_curve} object.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\description{
Plot method for infForest_curve objects
}
